Se cambia el tipo de dato que retorna

+++ Sofka.Piguinera.Cotizacion/Sofka.Piguinera.Cotizacion/Services/BookPricingService.cs
@@ -26,16 +26,16 @@ namespace Sofka.Piguinera.Cotizacion.Services
 
         public static void ApplyRetailIncrease(BaseBook book)
         {
-            const double RETAIL_INCREASE = 1.02;
+            const float RETAIL_INCREASE = 1.02f;
             book.CurrentPrice *= RETAIL_INCREASE;
         }
 
         private static  void ApplyBulkDecrease(BaseBook book)
         {
-            const double BULK_DECREASE_PER_UNIT = 0.15;
+            const float BULK_DECREASE_PER_UNIT = 0.15f;
             var discount = book.CurrentPrice * BULK_DECREASE_PER_UNIT;
 
-            book.Discount += (decimal)BULK_DECREASE_PER_UNIT;
+            book.Discount += (float)BULK_DECREASE_PER_UNIT;
             book.CurrentPrice -= discount;
         }
     }
